Fetch images concurrently
The URL lists for the images are quite old, so many of the images
are gone, and some of the servers don't even exist anymore. This
can cause the request to take a long time. Switch to asyncIO and
use several concurrent threads for fetching the images so that a
single time-consuming download doesn't block the whole process.
This speeds up the data fetching considerably.